requirements.txt
absl-py==2.1.0
acres==0.5.0
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.14.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.57.1
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
asyncssh==2.19.0
atpublic==5.0
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
boto3==1.38.27
botocore==1.38.27
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.0
catalogue==2.0.10
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.20.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.0
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.11
databricks-sdk==0.38.0
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.9
decorator==5.1.1
deeplake==3.9.47
defusedxml==0.7.1
Deprecated==1.2.15
dictdiffer==0.9.0
dill==0.4.0
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
dpath==2.2.0
dropbox==12.0.2
dulwich==0.22.7
dvc==3.58.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
dython==0.7.9
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
etelemetry==0.3.1
executing==2.1.0
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.16.1
filetype==1.2.0
findpython==0.6.3
firebase-admin==6.6.0
fitz==0.0.1.dev2
Flask==3.1.0
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.2
fqdn==1.5.1
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.12.0
funcy==2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.163.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googlemaps==4.10.0
gradio==5.36.2
gradio_client==1.10.4
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
groovy==0.1.2
grpcio==1.68.1
grpcio-status==1.68.1
gto==1.7.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.28.1
humbug==0.3.2
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
iterative-telemetry==0.0.9
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.7
kombu==5.4.2
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-deeplake==0.1.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.6
language_data==1.3.0
libclang==18.1.1
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.0
lz4==4.4.4
Mako==1.3.8
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mlflow==2.18.0
mlflow-skinny==2.18.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.18
murmurhash==1.0.12
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.16.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.5.1
omegaconf==2.3.0
openai==1.93.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatasets==0.1.22
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.13
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathlib==1.0.1
pathos==0.3.4
pathspec==0.12.1
pbs-installer==2025.5.17
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pillow_heif==0.22.0
pipreqs==0.5.0
pkginfo==1.12.1.2
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
poetry==2.1.3
poetry-core==2.1.3
pox==0.3.6
ppft==1.7.7
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
prov==2.1.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.3
pydub==0.25.1
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.0
pyproject_hooks==1.2.0
pytest==8.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roboflow==1.1.60
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.8
ruamel.yaml.clib==0.2.12
ruff==0.12.3
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.9
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.3
setuptools==75.6.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simplejson==3.20.1
six==1.16.0
sklearn-crfsuite==0.5.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.2
sqltrie==0.11.1
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
stone==3.3.1
streamlit==1.40.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.49.0
trove-classifiers==2025.5.9.12
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.95
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.31.2
voluptuous==0.15.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xattr==1.1.4
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
zstandard==0.23.0
